+++ plugins/record-hook/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @flatfile/plugin-record-hook
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [69fd41c]
+  - @flatfile/util-common@0.1.0
+
+## 1.0.2
+
+### Patch Changes
+
+- f5efd60: Update @flatfile/listener dependency
+
 ## 1.0.1
 
 ### Patch Changes