+++ utils/extractor/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @flatfile/plugin-extractor-utils
 
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [69fd41c]
+  - @flatfile/util-common@0.1.0
+
+## 0.2.3
+
+### Patch Changes
+
+- f5efd60: Update @flatfile/listener dependency
+- Updated dependencies [f5efd60]
+  - @flatfile/util-file-buffer@0.0.4
+
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies [490e8f1]
+  - @flatfile/util-file-buffer@0.0.3
+
 ## 0.2.1
 
 ### Patch Changes